+++ src/Admin/Field/LookupSlotField.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace App\Admin\Field;
+
+use App\Form\LookupSlotType;
+use EasyCorp\Bundle\EasyAdminBundle\Contracts\Field\FieldInterface;
+use EasyCorp\Bundle\EasyAdminBundle\Field\FieldTrait;
+use Symfony\Contracts\Translation\TranslatableInterface;
+
+final class LookupSlotField implements FieldInterface
+{
+    use FieldTrait;
+
+    /**
+     * @param TranslatableInterface|string|false|null $label
+     */
+    public static function new(string $propertyName, $label = null): self
+    {
+        return new self()
+            ->setProperty($propertyName)
+            ->setLabel($label)
+            ->setTemplatePath('admin/field/lookup_slot.html.twig')
+            ->setFormType(LookupSlotType::class)
+            ->addCssClass('lookup-slot');
+    }
+}