+++ src/Settings.php
@@ -43,7 +43,6 @@ final class Settings
             'enable_log_out' => t('Enable log out'),
             'category.user' => t('User'),
             'category.site' => t('Site'),
-            'max_loookups_per_day' => t('Max lookups per day'),
             'user_manual_url' => t('User manual URL'),
             'front_page_text' => t('Front page text'),
             'entry_expires_after' => t('Entry expires after'),